Revert "01.08"
This reverts commit 83ebd44696eb68eb206ce8e3aa832424b0382a94.

+++ src/main/java/ru/yandex/practicum/filmorate/controller/ControllerFilm.java
@@ -11,12 +11,15 @@ import ru.yandex.practicum.filmorate.service.ServiceFilm;
 import java.util.ArrayList;
 import java.util.List;
 
+
 @Slf4j
 @RestController
 @RequestMapping("/films")
 @RequiredArgsConstructor
 public class ControllerFilm {
+
 	private final ServiceFilm serviceFilm;
+
 
 	@GetMapping
 	public List<Film> getListOfFilms() {
@@ -37,18 +40,20 @@ public class ControllerFilm {
 	}
 
 	@GetMapping("/{filmId}")
-	public ResponseEntity<Film> lookForFilm(@PathVariable int filmId) {
+	public ResponseEntity<Film> lookForFilm(@PathVariable long filmId) {
 
 		return ResponseEntity.status(HttpStatus.OK).body(serviceFilm.getOfIdFilm(filmId));
 	}
 
+
 	@PutMapping("/{id}/like/{userId}")
-	public ResponseEntity<Film> favouriteFilms(@PathVariable int id, @PathVariable int userId) {
+	public ResponseEntity<Film> favouriteFilms(@PathVariable long id, @PathVariable long userId) {
 		return ResponseEntity.status(HttpStatus.OK).body(serviceFilm.joinLikeToFilm(id, userId));
 	}
 
+
 	@DeleteMapping("/{id}/like/{userId}")
-	public ResponseEntity<Film> deleteLikeInFilm(@PathVariable int id, @PathVariable int userId) {
+	public ResponseEntity<Film> deleteLikeInFilm(@PathVariable long id, @PathVariable long userId) {
 		return ResponseEntity.status(HttpStatus.OK).body(serviceFilm.deleteToLike(id, userId));
 	}
